requirements.txt
requests
beautifulsoup4
lxml
sqlalchemy
psycopg2
PyYAML